Try finding a subclass of given game protocol matching provided subsystem (in options)
This subsystem can be used as separate core/protocol class for different query protocols such as LAN. In some games an online server provides the query protocol differently than a local LAN server. This is the case for various Unreal Engine 3 games. In such case the subsystem can be provided by the option "subsystem". In case of using a "protocol-" type for querying, the variable is also passed as subsystem to find the specific protocol definition.